+++ test/language/array_types/cpp/PackedVariableArrayStructTest.cpp
@@ -46,12 +46,12 @@ protected:
         testStructure.setTestEnum(index % 2 == 0 ? TestEnum::DARK_RED : TestEnum::DARK_GREEN);
         testStructure.setTestBitmask(index % 2 == 0 ? TestBitmask::Values::READ : TestBitmask::Values::CREATE);
         if (index % 2 == 0)
-            testStructure.setTestOptional(index);
+            testStructure.setTestOptional(static_cast<uint16_t>(index));
         testStructure.setTestDynamicBitfield(index % 3);
         vector_type<uint64_t> values;
         for (uint64_t value = 1; value < 18; value += 3)
             values.push_back(value);
-        testStructure.setNumValues(values.size());
+        testStructure.setNumValues(static_cast<uint32_t>(values.size()));
         testStructure.setUnpackedValues(values);
         testStructure.setPackedValues(values);
 
@@ -62,7 +62,7 @@ protected:
     {
         TestChoice testChoice;
         if (index == 0 || index == 2 || index == 4)
-            testChoice.setValue16(index);
+            testChoice.setValue16(static_cast<uint16_t>(index));
         else if (index == 5)
             testChoice.setArray32(vector_type<uint32_t>{index * 2, index * 2 + 1});
         else
@@ -75,7 +75,7 @@ protected:
     {
         TestUnion testUnion;
         if (index % 2 == 0)
-            testUnion.setValue16(index);
+            testUnion.setValue16(static_cast<uint16_t>(index));
         else if (index == 5)
             testUnion.setArray32(vector_type<uint32_t>{index * 2, index * 2 + 1});
         else
@@ -84,23 +84,24 @@ protected:
         return testUnion;
     }
 
-    void checkBitSizeOf(size_t numElements)
+    void checkBitSizeOf(uint32_t numElements)
     {
         PackedVariableArray packedVariableArray;
         fillPackedVariableArray(packedVariableArray, numElements);
         packedVariableArray.initializeChildren();
 
-        const size_t unpackedBitSize = packedVariableArray.getTestUnpackedArray().bitSizeOf();
-        const size_t packedBitSize = packedVariableArray.getTestPackedArray().bitSizeOf();
+        const double unpackedBitSize = static_cast<double>(
+                packedVariableArray.getTestUnpackedArray().bitSizeOf());
+        const double packedBitSize = static_cast<double>(packedVariableArray.getTestPackedArray().bitSizeOf());
         const double minCompressionRatio = 0.59;
         ASSERT_GT(unpackedBitSize * minCompressionRatio, packedBitSize)
                 << "Unpacked array has " << std::to_string(unpackedBitSize) << " bits, "
                 << "packed array has " << std::to_string(packedBitSize) << " bits, "
                 << "compression ratio is "
-                << std::to_string(static_cast<double>(packedBitSize) /unpackedBitSize * 100) << "%!";
+                << std::to_string(packedBitSize /unpackedBitSize * 100) << "%!";
     }
 
-    void checkWriteRead(size_t numElements)
+    void checkWriteRead(uint32_t numElements)
     {
         PackedVariableArray packedVariableArray;
         fillPackedVariableArray(packedVariableArray, numElements);
